wrap <App> inside store provider
Give access to the state store to the <App> component, by
wrapping it inside <Provider> component.

Move all the code for <App> component to it's own module, to
make the wrapping and binding to the store easier.
